backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
pillow==11.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
firebase-admin==7.1.0
django-cors-headers==4.9.0
